:recycle: #852 - Refactor search form.
